Fix: 修复Computer初始化失败的bug；另外 cpu温度改为浮点数

+++ LibreHardwareMonitorLib/Warpper/StatusClasses.cs
@@ -50,10 +50,10 @@ public class StatusClasses
     public class CpuStatus
     {
         public int load;
-        public int temperature;
+        public float temperature;
         public int fan;
 
-        public CpuStatus(int load, int temperature, int fan)
+        public CpuStatus(int load, float temperature, int fan)
         {
             this.load = load;
             this.temperature = temperature;
